Move `commit_*_circuit` to separate module
This change moves the helper functions used to commit a 2- and 3- party
circuit on a Network that is already running into a separate module.
This allows the tests to use these functions within the `tests::admin`
module, allowing the tests to be separated more clearly based on
concern.
