+++ splinterd/tests/admin/mod.rs
@@ -15,6 +15,8 @@
 //! Admin service integration tests.
 
 #[cfg(feature = "node")]
+mod circuit_commit;
+#[cfg(feature = "node")]
 mod circuit_lifecycle;
 #[cfg(feature = "node")]
 mod circuit_list;
